% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.authorizationHeaderBearer}
\alias{.authorizationHeaderBearer}
\title{Constructs the value for the Bearer Authorization header}
\usage{
.authorizationHeaderBearer(jwt)
}
\description{
Constructs the value for the Bearer Authorization header
}
\keyword{internal}
