% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.authorizationHeaderRockr}
\alias{.authorizationHeaderRockr}
\title{Constructs the value for the Opal Authorization header}
\usage{
.authorizationHeaderRockr(username, password)
}
\value{
The header string
}
\description{
Constructs the value for the Opal Authorization header
}
\keyword{internal}
