% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.handleAttachment}
\alias{.handleAttachment}
\title{Handle the file content.}
\usage{
.handleAttachment(conn, response, disposition)
}
\value{
The response content object
}
\description{
Handle the file content.
}
\keyword{internal}
