% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.handleContent}
\alias{.handleContent}
\title{Handle the response content, unserialize when necessary.}
\usage{
.handleContent(conn, response)
}
\value{
The response content object
}
\description{
Handle the response content, unserialize when necessary.
}
\keyword{internal}
