% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.handleResponse}
\alias{.handleResponse}
\title{Default request response handler.}
\usage{
.handleResponse(conn, response)
}
\value{
The response content object
}
\description{
Default request response handler.
}
\keyword{internal}
