% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{.handleResponseOrCallback}
\alias{.handleResponseOrCallback}
\title{Process response with default handler or the provided one}
\usage{
.handleResponseOrCallback(conn, response, callback = NULL)
}
\value{
The response content object
}
\description{
Process response with default handler or the provided one
}
\keyword{internal}
