% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{rockr.close}
\alias{rockr.close}
\title{Close the R session}
\usage{
rockr.close(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\value{
Void
}
\description{
Close the R session, if there is any associated to the connection.
}
\examples{
\dontrun{
conn <- rockr.connect('administrator','password', url='https://rocker-demo.obiba.org')
rockr.open(conn)
rockr.close(conn)
}
}
\seealso{
Other connection functions: 
\code{\link{rockr.connect}()},
\code{\link{rockr.open}()}
}
\concept{connection functions}
