% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{rockr.command_rm}
\alias{rockr.command_rm}
\title{Remove an asynchronous command}
\usage{
rockr.command_rm(conn, id)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{id}{R command ID.}
}
\description{
Remove an asynchronous R commands in the remote R session.
}
\examples{
\dontrun{
o <- rockr.login('administrator','password', url='https://rocker-demo.obiba.org')
rockr.command_rm(o, '1234')
rockr.logout(o)
}
}
\seealso{
Other command functions: 
\code{\link{rockr.command_result}()},
\code{\link{rockr.commands_rm}()},
\code{\link{rockr.commands}()},
\code{\link{rockr.command}()}
}
\concept{command functions}
