% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{rockr.commands_rm}
\alias{rockr.commands_rm}
\title{Remove all asynchronous commands}
\usage{
rockr.commands_rm(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\description{
Remove all asynchronous R commands in the remote R session.
}
\examples{
\dontrun{
o <- rockr.login('administrator','password', url='https://rocker-demo.obiba.org')
rockr.commands_rm(o)
rockr.logout(o)
}
}
\seealso{
Other command functions: 
\code{\link{rockr.command_result}()},
\code{\link{rockr.command_rm}()},
\code{\link{rockr.commands}()},
\code{\link{rockr.command}()}
}
\concept{command functions}
