% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rockr.delete}
\alias{rockr.delete}
\title{Generic REST resource deletion.}
\usage{
rockr.delete(conn, ..., query = list(), callback = NULL)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{...}{Resource path segments.}

\item{query}{Named list of query parameters.}

\item{callback}{A callback function to handle the response object.}
}
\description{
Generic REST resource deletion.
}
\examples{
\dontrun{
conn <- rockr.login('administrator','password', url='https://rocker-demo.obiba.org')
rockr.delete(conn, 'some', 'resource')
rockr.logout(conn)
}
}
\seealso{
Other REST functions: 
\code{\link{rockr.get}()},
\code{\link{rockr.post}()},
\code{\link{rockr.put}()}
}
\concept{REST functions}
