% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{rockr.eval}
\alias{rockr.eval}
\title{Evaluate a R script}
\usage{
rockr.eval(conn, script, async = FALSE)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{script}{R script to execute.}

\item{async}{R script is executed asynchronously within the session (default is FALSE).
If TRUE, the value returned is the ID of the command to look for (from Opal 2.1).}
}
\description{
Execute a R script in the remote R session.
}
\examples{
\dontrun{
conn <- rockr.login(url='https://rocker-demo.obiba.org')
rockr.eval(conn, "x <- 'foo'")
rockr.eval(o, "ls()")
rockr.logout(conn)
}
}
\seealso{
Other execution functions: 
\code{\link{rockr.assign}()}
}
\concept{execution functions}
