% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rockr.get}
\alias{rockr.get}
\title{Generic REST resource getter.}
\usage{
rockr.get(
  conn,
  ...,
  query = list(),
  acceptType = "application/octet-stream, application/json",
  callback = NULL
)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{...}{Resource path segments.}

\item{query}{Named list of query parameters.}

\item{callback}{A callback function to handle the response object.}
}
\description{
Generic REST resource getter.
}
\examples{
\dontrun{
conn <- rockr.connect('administrator','password', url='https://rocker-demo.obiba.org')
rockr.open(conn)
rockr.get(conn, 'r', 'session', conn$session$id)
rockr.close(conn)
}
}
\seealso{
Other REST functions: 
\code{\link{rockr.delete}()},
\code{\link{rockr.post}()},
\code{\link{rockr.put}()}
}
\concept{REST functions}
\keyword{internal}
