% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rockr.log}
\alias{rockr.log}
\title{Log of the R server}
\usage{
rockr.log(conn, limit = 100)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{limit}{The number of lines in the tail.}
}
\value{
A vector of character strings, one per log file line.
}
\description{
Get the tail of the R server log.
}
\examples{
\dontrun{
conn <- rockr.connect(username='administrator', password='password',
                      url='https://rocker-demo.obiba.org')
rockr.log(conn, 10)
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.package_install}()},
\code{\link{rockr.package_rm}()},
\code{\link{rockr.packages_datashield}()},
\code{\link{rockr.packages_rm}()},
\code{\link{rockr.packages}()},
\code{\link{rockr.package}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.session_switch}()},
\code{\link{rockr.sessions}()},
\code{\link{rockr.session}()},
\code{\link{rockr.start}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()},
\code{\link{rockr.version}()}
}
\concept{administration functions}
