% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rockr.login}
\alias{rockr.login}
\title{Rocker login}
\usage{
rockr.login(username, password, token, url, opts = list(), restore = NULL)
}
\arguments{
\item{username}{User name in Rocker R server. Can be provided by "rockr.username" option.}

\item{password}{User password in Rocker R server. Can be provided by "rockr.password" option.}

\item{token}{Personal access token. Only effective if the username or the password is NULL or empty.
Can be provided by "rockr.token" option.}

\item{url}{Rocker R server url. Can be provided by "rockr.url" option.}

\item{opts}{Curl options as described by httr (call httr::httr_options() for details). Can be provided by "rockr.opts" option.}

\item{restore}{Workspace ID to be restored (see also rockr.logout)}
}
\value{
A rockr connection object.
}
\description{
Log in Rocker.
}
\examples{
\dontrun{
#### The below examples illustrate the different ways to login in rockr ####

# explicite username/password login
conn <- rockr.login(username='administrator', password='password',
                    url='https://rocker-demo.obiba.org')
rockr.logout(conn)

 # explicite personal access token login
conn <- rockr.login(token='HYG16LO0VaX4O0UardNbiqmr2ByBpRke',
                    url='https://rocker-demo.obiba.org')
rockr.logout(conn)

# login using options and user credentials
options(rockr.username='administrator',
 rockr.password='password',
 rockr.url='https://rocker-demo.obiba.org')
conn <- rockr.login()
rockr.logout(conn)

# login using options and personal access token
options(rockr.token='HYG16LO0VaX4O0UardNbiqmr2ByBpRke',
 rockr.url='https://rocker-demo.obiba.org')
conn <- rockr.login()
rockr.logout(conn)

# login using ssl key pair
options(rockr.opts=list(
   sslcert='my-publickey.pem',
   sslkey='my-privatekey.pem'))
conn <- rockr.login(url='https://rocker-demo.obiba.org')
rockr.logout(conn)
}
}
\seealso{
Other connection functions: 
\code{\link{rockr.logout}()}
}
\concept{connection functions}
