% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rockr.logout}
\alias{rockr.logout}
\title{Logout from Rocker}
\usage{
rockr.logout(conn, save = FALSE)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{save}{Save the workspace with given identifier (default value is FALSE, current session ID if TRUE).}
}
\description{
Clear the R sessions and logout from Rocker R server.
}
\examples{
\dontrun{
conn <- rockr.login('administrator','password', url='https://rocker-demo.obiba.org')
rockr.logout(conn)
}
}
\seealso{
Other connection functions: 
\code{\link{rockr.login}()}
}
\concept{connection functions}
