% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{rockr.packages_datashield}
\alias{rockr.packages_datashield}
\title{List of the settings of the DataSHIELD packages}
\usage{
rockr.packages_datashield(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\value{
The list of the DataSHIELD R packages properties (AggregateMethods, AssignMethods, Options)
}
\description{
Discover the list of DataSHIELD packages and their settings from the R server.
}
\examples{
\dontrun{
conn <- rockr.connect(username='administrator', password='password',
                      url='https://rocker-demo.obiba.org')
rockr.packages_datashield(conn)
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.log}()},
\code{\link{rockr.package_install}()},
\code{\link{rockr.package_rm}()},
\code{\link{rockr.packages_rm}()},
\code{\link{rockr.packages}()},
\code{\link{rockr.package}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.session_switch}()},
\code{\link{rockr.sessions}()},
\code{\link{rockr.session}()},
\code{\link{rockr.start}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()},
\code{\link{rockr.version}()}
}
\concept{administration functions}
