% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{rockr.packages_rm}
\alias{rockr.packages_rm}
\title{Remove some packages from the R server}
\usage{
rockr.packages_rm(conn, names)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{names}{The package names.}
}
\description{
Remove some packages from the R server.
}
\examples{
\dontrun{
conn <- rockr.connect(username='administrator', password='password',
                      url='https://rocker-demo.obiba.org')
rockr.packages_rm(conn, c('rlang', 'tibble'))
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.log}()},
\code{\link{rockr.package_install}()},
\code{\link{rockr.package_rm}()},
\code{\link{rockr.packages_datashield}()},
\code{\link{rockr.packages}()},
\code{\link{rockr.package}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.session_switch}()},
\code{\link{rockr.sessions}()},
\code{\link{rockr.session}()},
\code{\link{rockr.start}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()},
\code{\link{rockr.version}()}
}
\concept{administration functions}
