% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rockr.post}
\alias{rockr.post}
\title{Generic REST resource creation.}
\usage{
rockr.post(
  conn,
  ...,
  query = list(),
  body = "",
  contentType = "application/x-rscript",
  callback = NULL
)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{...}{Resource path segments.}

\item{query}{Named list of query parameters.}

\item{body}{The body of the request.}

\item{contentType}{The type of the body content.}

\item{callback}{A callback function to handle the response object.}
}
\description{
Generic REST resource creation.
}
\examples{
\dontrun{
conn <- rockr.login('administrator','password', url='https://rocker-demo.obiba.org')
rockr.post(conn, 'some', 'resources', body='{"some":"value"}')
rockr.logout(conn)
}
}
\seealso{
Other REST functions: 
\code{\link{rockr.delete}()},
\code{\link{rockr.get}()},
\code{\link{rockr.put}()}
}
\concept{REST functions}
