% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{rockr.put}
\alias{rockr.put}
\title{Generic REST resource update.}
\usage{
rockr.put(
  conn,
  ...,
  query = list(),
  body = "",
  contentType = "application/json",
  callback = NULL
)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{...}{Resource path segments.}

\item{query}{Named list of query parameters.}

\item{body}{The body of the request.}

\item{contentType}{The type of the body content. Default is 'application/json'.}

\item{callback}{A callback function to handle the response object.}
}
\description{
Generic REST resource update.
}
\examples{
\dontrun{
conn <- rockr.login('administrator','password', url='https://rocker-demo.obiba.org')
rockr.put(conn, 'some', 'resource', 'toupdate', body='{"some":"value"}')
rockr.logout(conn)
}
}
\seealso{
Other REST functions: 
\code{\link{rockr.delete}()},
\code{\link{rockr.get}()},
\code{\link{rockr.post}()}
}
\concept{REST functions}
