% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rockr.start}
\alias{rockr.start}
\title{Start the R server}
\usage{
rockr.start(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\value{
Logical indicating wether the R server is running (invisible).
}
\description{
Start the R server, if not already running.
}
\examples{
\dontrun{
conn <- rockr.connect(username='administrator', password='password',
                      url='https://rocker-demo.obiba.org')
rockr.start(conn)
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.log}()},
\code{\link{rockr.package_install}()},
\code{\link{rockr.package_rm}()},
\code{\link{rockr.packages_datashield}()},
\code{\link{rockr.packages_rm}()},
\code{\link{rockr.packages}()},
\code{\link{rockr.package}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.session_switch}()},
\code{\link{rockr.sessions}()},
\code{\link{rockr.session}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()},
\code{\link{rockr.version}()}
}
\concept{administration functions}
