% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rockr.status}
\alias{rockr.status}
\title{Status of the R server}
\usage{
rockr.status(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\value{
The R server status object.
}
\description{
Get the status of the R server, including information about the server node.
}
\examples{
\dontrun{
conn <- rockr.connect(username='administrator', password='password',
                      url='https://rocker-demo.obiba.org')
rockr.status(conn)
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.log}()},
\code{\link{rockr.package_install}()},
\code{\link{rockr.package_rm}()},
\code{\link{rockr.packages_datashield}()},
\code{\link{rockr.packages_rm}()},
\code{\link{rockr.packages}()},
\code{\link{rockr.package}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.session_switch}()},
\code{\link{rockr.sessions}()},
\code{\link{rockr.session}()},
\code{\link{rockr.start}()},
\code{\link{rockr.stop}()},
\code{\link{rockr.version}()}
}
\concept{administration functions}
