% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{rockr.version}
\alias{rockr.version}
\title{Version of the R server}
\usage{
rockr.version(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\description{
Get the version object of the R server.
}
\examples{
\dontrun{
conn <- rockr.connect(url='https://rocker-demo.obiba.org')
rockr.version(conn, 10)
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.log}()},
\code{\link{rockr.package_install}()},
\code{\link{rockr.package_rm}()},
\code{\link{rockr.packages_datashield}()},
\code{\link{rockr.packages}()},
\code{\link{rockr.package}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.start}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()}
}
\concept{administration functions}
