% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{rockr.assign}
\alias{rockr.assign}
\title{Evaluate a R script}
\usage{
rockr.assign(conn, symbol, value, async = FALSE)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{symbol}{Name of the R symbol.}

\item{value}{Value to assign to the symbol: can be a raw value or a R script to execute.}

\item{async}{R script is executed asynchronously within the session (default is FALSE).
If TRUE, the value returned is the ID of the command to look for.}
}
\description{
Execute a R script in the remote R session.
}
\examples{
\dontrun{
conn <- rockr.connect(url='https://rocker-demo.obiba.org')
rockr.open(conn)
rockr.eval(conn, "x <- 'foo'")
rockr.eval(conn, "ls()")
rockr.close(conn)
}
}
\seealso{
Other execution functions: 
\code{\link{rockr.eval}()}
}
\concept{execution functions}
