% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{rockr.package}
\alias{rockr.package}
\title{Package description from the R server}
\usage{
rockr.package(conn, name)
}
\arguments{
\item{conn}{A rockr connection object.}

\item{name}{The package name.}
}
\value{
The R package description object.
}
\description{
Get the package description from the R server.
}
\examples{
\dontrun{
conn <- rockr.connect(username='administrator', password='password',
                      url='https://rock-demo.obiba.org')
rockr.package(conn, 'tibble')
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.log}()},
\code{\link{rockr.package_install}()},
\code{\link{rockr.package_rm}()},
\code{\link{rockr.packages_datashield}()},
\code{\link{rockr.packages_rm}()},
\code{\link{rockr.packages}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.session_switch}()},
\code{\link{rockr.sessions}()},
\code{\link{rockr.session}()},
\code{\link{rockr.start}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()},
\code{\link{rockr.version}()}
}
\concept{administration functions}
