% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{rockr.packages}
\alias{rockr.packages}
\title{List of the packages in the R server}
\usage{
rockr.packages(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\description{
Get the list of packages from the R server.
}
\examples{
\dontrun{
conn <- rockr.connect(url='https://rocker-demo.obiba.org')
rockr.packages(conn)
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.restart}()},
\code{\link{rockr.start}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()}
}
\concept{administration functions}
