% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{rockr.start}
\alias{rockr.start}
\title{Start the R server}
\usage{
rockr.start(conn)
}
\arguments{
\item{conn}{A rockr connection object.}
}
\value{
Logical indicating wether the R server is running (invisible).
}
\description{
Start the R server, if not already running.
}
\examples{
\dontrun{
conn <- rockr.connect(url='https://rocker-demo.obiba.org')
rockr.start(conn)
}
}
\seealso{
Other administration functions: 
\code{\link{rockr.packages}()},
\code{\link{rockr.restart}()},
\code{\link{rockr.status}()},
\code{\link{rockr.stop}()}
}
\concept{administration functions}
